% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entry_functions.R
\name{get_entry}
\alias{get_entry}
\title{get_entry}
\usage{
get_entry(entryid = NULL)
}
\arguments{
\item{entryid}{The entry ID(s). Can be found on the FPL website under 'Gameweek history' in the URL - https://fantasy.premierleague.com/a/entry/XXXXXX/history.}
}
\description{
This function fetches detailed information for a Fantasy Premier League entry given the entry ID(s).
}
\examples{
get_entry(1076)
}
\keyword{entry}
