% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entry_functions.R
\name{get_entry_season}
\alias{get_entry_season}
\title{get_entry_season}
\usage{
get_entry_season(entryid = NULL)
}
\arguments{
\item{entryid}{The entry ID(s). Can be found on the FPL website under 'Gameweek history' in the URL - https://fantasy.premierleague.com/a/entry/XXXXXX/history. You can list multiple, e.g. with c().}
}
\description{
This function fetches in-season statistics for a Fantasy Premier League entry given the entry ID(s).
}
\examples{
get_entry_season(1076)
}
\keyword{entry}
