% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_functions.R
\name{get_fdr}
\alias{get_fdr}
\title{get_fdr}
\usage{
get_fdr()
}
\description{
This function fetches FPL's 'team strength' measures, which form the basis of the 'FDR' (Fixture Difficulty Rating)
}
\examples{
get_fdr()
}
\keyword{teams}
