% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/league_functions.R
\name{get_league}
\alias{get_league}
\title{get_league}
\usage{
get_league(leagueid = NULL, leaguetype = "classic")
}
\arguments{
\item{leagueid}{The league ID(s). Can be found on the FPL website.}

\item{leaguetype}{The league type: 'classic' or 'h2h'.}
}
\description{
This function fetches statistics for a Fantasy Premier League mini-league given the league ID and type. IMPORTANT NOTE: FPL generally asks scrapers NOT to abuse the /league sites (see: https://fantasy.premierleague.com/robots.txt).
}
\examples{
get_league(441,"classic")
}
\keyword{league}
