% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/league_functions.R
\name{get_league_entries}
\alias{get_league_entries}
\title{get_league_entries}
\usage{
get_league_entries(leagueid = NULL, leaguetype = "classic", pages = 1)
}
\arguments{
\item{leagueid}{The league ID(s). Can be found on the FPL website.}

\item{leaguetype}{The league type: 'classic' or 'h2h'.}

\item{pages}{The number of pages of entries to fetch (there are 50 entries on each page)}
}
\description{
This function fetches a list of entries in the standings for a Fantasy Premier League mini-league given the league ID and type. In the new FPL API for 19/20, this information requires authentication - so you'll be prompted to post your FPL login and password to retrieve the data.
}
\keyword{league}
