% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/player_functions.R
\name{get_player_details}
\alias{get_player_details}
\title{get_player_details}
\usage{
get_player_details(playerid = NULL)
}
\arguments{
\item{playerid}{The player's ID. Can be found using get_player_id(). You can list multiple , e.g. with c(). If blank, the function fetches all current players - note that takes a while to load.}
}
\description{
This function fetches detailed, gameweek-by-gameweek information for the current season on selected players given player ID(s).
}
\examples{
get_player_details(300)
}
\keyword{player}
