% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/player_functions.R
\name{get_player_hist}
\alias{get_player_hist}
\title{get_player_hist}
\usage{
get_player_hist(playerid = NULL)
}
\arguments{
\item{playerid}{The player's ID. Can be found using get_player_id(). You can list multiple , e.g. with c(). If blank, the function fetches all current players - note that takes a while to load.}
}
\description{
This function fetches historical summary information on selected players given player ID(s).
}
\examples{
get_player_hist(playerid=200)

get_player_hist(playerid=get_player_id("Sead Kolasinac")$id)
}
\keyword{player}
