% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/player_functions.R
\name{get_player_id}
\alias{get_player_id}
\title{get_player_id}
\usage{
get_player_id(name = NULL)
}
\arguments{
\item{name}{The player's full name, as listed on the official Fantasy Premier League site (for instance: "Richarlison de Andrade", not "Richarlison". You can list multiple, e.g. with c().}
}
\description{
This function fetches the FPL ID of a player given the player's full name. If blank, the function fetches all current players.
}
\examples{
get_player_id(name="Virgil van Dijk")
}
\keyword{player}
