% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/player_functions.R
\name{get_player_ids}
\alias{get_player_ids}
\title{get_player_ids}
\usage{
get_player_ids(season = NULL)
}
\arguments{
\item{season}{To retrieve prior season data, enter a 2-digit year corresponding to the start of the historical FPL season of interest (e.g. '17' for the 2017/2018 season).}
}
\description{
This function fetches ids for all players.
}
\examples{
get_player_ids()
}
\keyword{player}
