% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/player_functions.R
\name{get_player_info}
\alias{get_player_info}
\title{get_player_info}
\usage{
get_player_info(name = NULL)
}
\arguments{
\item{name}{The player's full name, as listed on the official Fantasy Premier League site (for instance: "Richarlison de Andrade", not "Richarlison"). If blank, the function fetches all players.}
}
\description{
This function fetches selected summary information on selected players given full player name(s).
}
\examples{
get_player_info(name=c("Jesse Lingard","Aleksandar Mitrovic"))
}
\keyword{player}
