% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/player_functions.R
\name{get_player_name}
\alias{get_player_name}
\title{get_player_name}
\usage{
get_player_name(playerid = NULL)
}
\arguments{
\item{playerid}{The player's FPL ID. You can list multiple, e.g. with c().}
}
\description{
This function fetches a player's full name given FPL ID. If blank, the function fetches all current players.
}
\examples{
get_player_name(playerid=300)

get_player_name(playerid=150:160)
}
\keyword{player}
