% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/round_functions.R
\name{get_round_info}
\alias{get_round_info}
\title{get_round_info}
\usage{
get_round_info(round = NULL)
}
\arguments{
\item{round}{The FPL Gameweek number for which basic summary information is requested, e.g. 20. You can list multiple rounds, e.g. with c(). If blank, the function fetches all Gameweeks.}
}
\description{
This function fetches basic summary information on FPL Gameweeks.
}
\examples{
get_round_info(2:4)
}
\keyword{round}
