% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fplscrapR.R
\name{getPlayerID}
\alias{getPlayerID}
\title{getPlayerID}
\usage{
getPlayerID(name = NULL)
}
\arguments{
\item{name}{The player's full name, as listed on the official Fantasy Premier League site (for instance: "Richarlison de Andrade", not "Richarlison". You can list multiple rounds, e.g. with c().}
}
\description{
This function fetches the FPL ID of a player given the player's full name.
}
\examples{
getPlayerID()
}
\keyword{player}
