% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fplscrapR.R
\name{getPlayerInfo}
\alias{getPlayerInfo}
\title{getPlayerInfo}
\usage{
getPlayerInfo(name = NULL)
}
\arguments{
\item{name}{The player's full name, as listed on the official Fantasy Premier League site (for instance: "Richarlison de Andrade", not "Richarlison"). If blank, the function fetches all players.}
}
\description{
This function fetches selected summary information on selected players given full player name(s).
}
\examples{
getPlayerInfo()
}
\keyword{player}
