% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fplscrapR.R
\name{getPlayerName}
\alias{getPlayerName}
\title{getPlayerName}
\usage{
getPlayerName(playerid = NULL)
}
\arguments{
\item{playerid}{The player's FPL ID. You can list multiple rounds, e.g. with c().}
}
\description{
This function fetches a player's full name given FPL ID.
}
\examples{
getPlayerName()
}
\keyword{player}
