% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entry_functions.R
\name{get_entry_player_picks}
\alias{get_entry_player_picks}
\title{get_entry_player_picks}
\usage{
get_entry_player_picks(entryid = NULL, gw = NULL)
}
\arguments{
\item{entryid}{The entry ID(s). Can be found on the FPL website under 'Gameweek history' in the URL - https://fantasy.premierleague.com/a/entry/XXXXXX/history.}

\item{gw}{The GW(s) for which player picks is requested.}
}
\value{
A data.frame of player picks for the requested entry ID(s) and GW(s).
}
\description{
This function fetches detailed information for a Fantasy Premier League entry given the entry ID(s) and GW(s).
}
\examples{

get_entry_player_picks(c(1076, 1077), 1:10)
}
\keyword{entry}
