% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/league_functions.R
\name{get_league}
\alias{get_league}
\title{get_league}
\usage{
get_league(leagueid = NULL, leaguetype = "classic")
}
\arguments{
\item{leagueid}{The league ID(s). Can be found on the FPL website.}

\item{leaguetype}{The league type: 'classic' or 'h2h'.}
}
\description{
This function fetches basic information for a Fantasy Premier League mini-league given the league ID and type. In the new FPL API for 19/20, this information requires authentication - so you'll be prompted to post your FPL login and password to retrieve the data.
}
\keyword{league}
